requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
cssselect2==0.8.0
Django==5.1.2
django-bootstrap5==24.3
django-cors-headers==4.7.0
django-debug-toolbar==4.4.6
django-elasticsearch-dsl==8.0
django-js-asset==3.0.1
django-mptt==0.16.0
django-simple-history==3.10.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
elastic-transport==8.17.1
elasticsearch==8.18.0
elasticsearch-dsl==8.18.0
fonttools==4.59.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.6
pdf2image==1.17.0
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
